My first instinct should be to go into Ship.php and copy all of the contents and paste that into RebelShip.php since most of it will probably apply. But I shouldn't need to remind you that this would be a silly amount of duplication in our code which would make everyone sad. This is our chance to let classes help us not be sad by using the extends keyword.
By saying class RebelShip extends Ship everything that's in the Ship class is automatically inside of RebelShip.

It's as if all the properties and methods of Ship are now a part of the RebelShip blueprint.

In index.php we can say {dollar_sign}rebelShip = new RebelShip('My new rebel ship'); and we can just add this to the  array.

Remember, down here we iterate over the ships and call things like getName(), getWeaponPower() and getJediFactor() which don't actually live inside of RebelShip:

 124 lines  index.php
... lines 1 - 72
  <?php foreach ({dollar_sign}ships as {dollar_sign}ship): ?>
    <tr>
      <td><?php echo {dollar_sign}ship->getName(); ?></td>
      <td><?php echo {dollar_sign}ship->getWeaponPower(); ?></td>
... lines 77 - 85
    </tr>
  <?php endforeach; ?>
... lines 88 - 124
But when we refresh, it works perfectly!

Lesson number 1: when you have one class that extends another, it inherits (you'll hear that word a lot) all of the stuff inside that parent class. So we can call methods like getName() or getNameAndSpecs() on RebelShip because it inherits that from Ship.