I want to create ship objects for the empire and the rebels. So let's do that, if ({dollar_sign}shipData['team'] == 'rebel') which is the key inside the database. Then we'll have {dollar_sign}ship = new RebelShip({dollar_sign}shipData['name']);. Else, we'll throw in our normal ship line, which represents the Empire ship.
Ok, this doesn't have anything to do with Object Oriented coding, it's just a nice example of a use case for multiple classes. We have a database table, and you can create different objects from that table. This is nice because we'll be able to have these two objects behave differently.