So far RebelShip and Ship have all the same stuff except for the one extra method I have on RebelShip that I'm not using.
If we go back and refresh, everything still works perfectly! Now, technically I'm fairly certain that two of these are RebelShip objects and two are Ship objects but we can't really tell right now. Clearly we need to add identifiers so we know who to cheer on.

To do this, start by adding public function getType() to our Ship and return a description, like 'Empire'.
Since we added that to Ship, we can call getType on both Ship and RebelShip.

Back to the browser and refresh. Everything has joined to fight for the Empire! Which makes sense. Both ship classes use this same method.

Time for the next really powerful thing with inheritance. In addition to adding methods to a sub class like RebelShip you can override methods. Copy the getType from Ship and paste it into RebelShip and change what it returns to 'Rebel'.

RebelShip copies the entire blue print of Ship but it can replace any of those pieces. When we refresh now, we have two 'Rebel' ships in addition to our two 'Empire' ships. Excellent!