A key part of this is that the parent getType class is never called for all rebel ship objects it is completely replaced. If I echo 'Parent Function' inside of getType in the Ship class and refresh, we see our ugly text echoing for the Empire ships and not the Rebel ships. This is thanks to our parent function not being called in RebelShip.
On to more methods, another one on Ship is isFunctional which we setup to have a 30% chance of a ship being broken, which is what our cute cloud here indicates:
lib/Model/Ship.php
... lines 1 - 2
class Ship
{
... lines 5 - 16
  public function __construct({dollar_sign}name)
  {
... lines 19 - 20
    ! {dollar_sign}this->underRepair = mt_rand(1, 100) < 30;
  }
  public function isFunctional()
  {
    return {dollar_sign}this->underRepair;
  }
... lines 28 - 138
}

But, we all know that the Rebels are really scrappy and they don't have the luxury of letting their ships get broken. Even if they are kinda broken they still fly and make it work. Which is just one more reason why the rebels are awesome.

So I need to set this up so the Rebel ships are never showing as broken which we can do really easily by overriding isFunctional inside of RebelShip. Let's update this to return true, which will never show a rebel ship as broken.

When we refresh now the Rebel ships always have sunshine, and the Empire ships sometimes have adorable clouds.

By having two classes we are starting to shape the different behaviors and properties of each, while still keeping most things in common and not duplicated.