Let's keep making our Rebel ships work a bit differently than the Empire's. In this dropdown you can see a short summary of each ship that is currently functional. It shows their name, weapon power, jedi power and strength which all comes from the getNameAndSpecs function. But I would like a way to tell which ships in this list align with the rebels, so let's add that word in parenthesis at the end.
As usual to do that, we'll override this in RebelShip. Copy the getNameAndSpecs function and paste it over here. And then just add '(rebel)' at the end.

Now you may be thinking 'guys, that's some serious code duplication...'. Well you're absolutely right, and we'll get to fixing that!

For now what we've got is pretty straightforward, so let's refresh and... oh, check out our dropdown. We've got an Undefinded property RebelShip::{dollar_sign}name error.

You can't access private things in sub-classes

Back in PhpStorm, you can see ->name is highlighted with an error message of 'Member has private access'. Interesting. So far, I've told you that since RebelShip extends Ship it has access to everything inside of it like the properties and methods as if they also exist inside of RebelShip. However, this error really does seem to be saying something different than that.

We can see that in Ship there is a name property so why isn't this working? The answer has to do with this word private in front of {dollar_sign}name:

 140 lines  lib/Model/Ship.php
... lines 1 - 2
class Ship
{
... lines 5 - 6
    private {dollar_sign}name;
... lines 8 - 138
}

All functions and properties so far are either private or public. If a function or a property is private it means you can only access it from within the ship class:

 140 lines  lib/Model/Ship.php
... lines 1 - 2
class Ship
{
... lines 5 - 16
  public function __construct({dollar_sign}name)
  {
    {dollar_sign}this->name = {dollar_sign}name;
... lines 20 - 21
  }
... lines 23 - 138
}

Like here where we say {dollar_sign}this->name. As we can see here, private functions and properties can't be accessed inside of subclasses. So only things inside of the Ship class can access this private ; property.

I always recommend that you make things private until you need to access them from outside of the class you're working in.