Let's go ahead and just change properties to methods in RebelShip.
I already have these public functions so why not use them? It allows me to keep these properties private which is looking ahead a little bit, but the more things you have marked as private the easier it's going to be to maintain and update your code later.

Back to the browser and refresh, and things still work!