Let's temporarily add a new private function to Ship called getSecretDoorCodeToTheDeathstar(). Since only Empire ships should have access to this you can see why setting it as private makes sense. And let's return the secret code 'Ra1nb0ws'
Over in RebelShip I should not be able to access this new function since we set it to private.

We see the 'Member has private access' error so when we refresh we can check the dropdown to confirm that things aren't working. Fatal error: Call to private method Ship::getSecretDoorCodeToTheDeathstar() and we need to view the source to see the full error message.