We covered that when you override a function, you override it entirely. In RebelShip we're overriding getNameAndSpecs:
 44 lines  lib/Model/RebelShip.php
... lines 1 - 2
class RebelShip extends Ship
{
... lines 5 - 22
    public function getNameAndSpecs({dollar_sign}useShortFormat = false)
    {
      if () {
        return sprintf(
          '%s: %s/%s/%s (Rebel)',
          {dollar_sign}this->getName(),
          {dollar_sign}this->getWeaponPower(),
          {dollar_sign}this->getJediFactor(),
          {dollar_sign}this->getStrength()
        );
      } else {
        return sprintf(
          '%s: w:%s, j:%s, s:%s (Rebel)',
          {dollar_sign}this->getName(),
          {dollar_sign}this->getWeaponPower(),
          {dollar_sign}this->getJediFactor(),
          {dollar_sign}this->getStrength()
        );
      }
    }
}
which means that when this method is called on a RebelShip object the getNameAndSpecs inside of the original Ship class, i.e. the parent class, is never called. In this case that's sort of a problem because it leaves us with all this code duplication. It would be way better if we could somehow call the parent method, getNameAndSpecs inside of Ship, and then just add this '(rebel)' part to the end.