We saw in the last chapter, that from within RebelShip you can call methods that exist in the parent class as long as they are public or protected. Let's try that here. Add {dollar_sign}val = {dollar_sign}this->getNameAndSpecs(). Pass in the  then {dollar_sign}val .= ('Rebel'); and finally return {dollar_sign}val.
Doesn't that look a whole lot nicer? Yes, yes it does.

Let's give our experiment here a try. Refresh! Hmmm something is wrong... (!) Fatal error: Maximum, let's view the source code since this error is stuck in our select box. Ah there we go:  Fatal error: Maximum function nesting level of '200' reached, aborting... lines 28 - 138 This means that we have a loop in our code, on index line 98 we call getNameAndSpecs and then on line 25 of RebelShip we call getNameAndSpecs again. This isn't working because when we call ->getNameAndSpecs, it's literally calling this same method again inside of RebelShip not the parent function in Ship.