Fat Classes Over in PhpStorm, when we look at this function now, Ship has a Jedi Factor property but RebelShip doesn't need that at all. Since RebelShip is extending Ship it is still inheriting that property. While this doesn't hurt anything it is a bit weird to have this extra property on our class that we aren't using at all. And this is also true for the isFunctional() method. In RebelShip it's always true:
 36 lines  lib/Model/RebelShip.php
... lines 1 - 2
class RebelShip extends Ship
{
... lines 5 - 17
    public function isFunctional()
    {
        return true;
    }
... lines 22 - 34
}
But in Ship it reads from an underRepair property, and again that's just not needed in RebelShip:

 140 lines  lib/Model/Ship.php
... lines 1 - 2
class Ship
{
... lines 5 - 14
    private {dollar_sign}underRepair;
... lines 16 - 23
    public function isFunctional()
    {
        return ! {dollar_sign}this->underRepair;
    }
... lines 28 - 138
}
The point being, Ship comes with extra stuff that we are inheriting but not using in RebelShip.

These classes are like blueprints, so maybe, instead of having RebelShip extend Ship and inherit all these things it won't use, we should have a third class that would hold the properties and methods that actually overlap between the two called AbstractShip. From here, Ship and RebelShip would both extend AbstractShip to get access to those common things.

This is a way of changing the class heirachy so that each class has only what it actually needs.