Let's trim down AbstractShip to only the items that are truly shared between our two ships.
First, jediFactor is specific to Ship so let's move it over there.

And then we'll update the references to it in AbstractShip to what the two classes share, which is a getJediFactor() function.

So let's copy and paste that function into Ship.

RebelShip already has one so that class is good to go already. Now in AbstractShip the getJediFactor() function will either call the version of the function in Ship or RebelShip depending on what is being loaded. There are a few other things I want to share with you about this, but we'll get to those later.