Since everything seems to be working on our site, let's start a battle! Four Jedi Starfighters against three Super Star Destroyers. Engage.
Ahh an error!

Argument 1 passed to BattleManager::battle() must be an instance of Ship, instance of RebelShip given
And this is apparently happening on battle line 32:

 109 lines  battle.php
... lines 1 - 33
{dollar_sign}battleResult = {dollar_sign}battleManager->battle({dollar_sign}ship1, {dollar_sign}ship1Quantity, {dollar_sign}ship2, {dollar_sign}ship2Quantity);
... lines 35 - 109
And BattleManager line 10:

 66 lines  lib/Service/BattleManager.php
... lines 1 - 2
class BattleManager
{
... lines 5 - 9
    public function battle(Ship {dollar_sign}ship1, {dollar_sign}ship1Quantity, Ship {dollar_sign}ship2, {dollar_sign}ship2Quantity)
    {
... lines 12 - 56
    }
... lines 58 - 64
}
Back to our IDE and open up battle.php.

Trouble With Type Hints

Down on line 32, what we see is that  is actually a RebelShip object, which makes sense since one of the ships I selected was a Rebel. But it expected that to be a normal Ship class. Over in BattleManager look at the battle function to see the problem! We type hinted our arguments with the Ship class:

 66 lines  lib/Service/BattleManager.php
... lines 1 - 2
class BattleManager
{
... lines 5 - 9
    public function battle(Ship {dollar_sign}ship1, {dollar_sign}ship1Quantity, Ship {dollar_sign}ship2, {dollar_sign}ship2Quantity)
    {
... lines 12 - 56
    }
... lines 58 - 64
}
Which tells PHP to only allow Ship classes or subclasses to be passed here.