What Methods are really on AbstractShip?
Now we have a few minor, but interesting, problems. First, in AbstractShip head down to getNameAndSpecs() and we see that getJediFactor() is highlighted with an error that says 'Method getJediFactor() not found in class AbstractShip'. Now, this is working because we do have a getJediFactor() method in Ship and RebelShip. When we call getNameAndSpecs() it's able to call getJediFactor(). But this should look a little fishy to you. There is no getJediFactor() function inside of AbstractShip, so just looking at this class you should feel suspicious and question whether or not this works.

Here's what's going on, we have an implied rule that says, 'Yo, every class that extends AbstractShip must have a getJediFactor() function.' If it doesn't everything is going to break when we call this function with a 'method not found' error. We aren't enforcing this rule. So we could easily create a new ship class, extend AbstractShip, and forget to add a getJediFactor() function. Our application would break and no battles would be happening. Sad times.