There's a feature called Abstract Classes that can handle this issue for us. I'll scroll up, but really the position of this doesn't matter. Add a new abstract public function getJediFactor(). You may notice there are two different things about this. One is the word abstract before public function and the other is that I just have a semicolon on the end, I didn't actually make a function. The best part, this line doesn't add any functionality to our app, but it does force any class that extends this to have this method.
For example, if RebelShip didn't have this getJediFactor() method, then when we refresh the browser we'll get a huge error that says: 'Hey! RebelShip must have a getJediFactor function!'. This is because it has been defined as an abstract function inside of the parent class.

Up until now we could have instantiated an abstract ship directly with new AbstractShip() we didn't actually want to but it was possible. But, once you have an abstract function in here, that is no longer an option, it's only purpose then becomes to be a blueprint for other classes to extend.

Marking a Class as Abstract

Up here at the top of the file you can see that there is an error highlight with a message that says 'Class must be declared abstract or implement method getJediFactor()'. Once your class has an abstract function you need to add the abstract keyword in front of it, which enforces the rule that you can't say new AbstractShip().
Now when we scroll down, we can see that getJediFactor() isn't highlighted anymore since we know that inside AbstractShip any subclasses will be forced to have that. Back to the browser and refresh! Everything still works just fine.