Related to this, there is one more little thing we need to fix up. Start in ShipLoader, notice that our getShips() and findOneById() functions still had PHPDoc above them that say they return a ship object. That's not the biggest deal, but it would be more accurate if it said AbstractShip - because this actually returns a mixture of RebelShip and Ship objects. Now check this out, inside of index.php, remember this  variable we get by calling that getShips() function?
 123 lines  index.php
... lines 1 - 6
{dollar_sign}ships = {dollar_sign}shipLoader->getShips();
... lines 8 - 123
So that returns an array of AbstractShip objects. When we loop over it, the isFunctional() and the getType() functions aren't found:

 123 lines  index.php
... lines 1 - 70
  <?php foreach ({dollar_sign}ships as {dollar_sign}ship): ?>
... lines 72 - 74
    <td><?php echo {dollar_sign}ship->getJediFactor(); ?></td>
... line 76
     <td><?php echo {dollar_sign}ship->getType(); ?></td>
... lines 78 - 85
  <?php endforeach; ?>

The message here says 'Method getType() not found in class AbstractShip'. This is just like the getJediFactor() problem we just fixed. We don't have a getType() function inside of here. Both of our subclasses do, which is why our app still works, but technically we're not enforcing that. Any new subclasses to AbstractShip could easily end up missing these functions which would again stop all the battles.