Any new subclasses to AbstractShip could easily end up missing these functions which would again stop all the battles. What we need is another abstract public function for getType() and isFunctional().
This doesn't change anything in our application, it just forces our subclasses to have those methods. And now index.php is really happy again!

That's the power of abstract classes, you can have a whole bunch of shared logic in there, but if there are a couple of pieces that you can't fill in in your abstract class because they are specific to your subclasses, no problem! Just put them in there as abstract functions and your subclasses will be forced to have those.

In my example these are abstract public functions but you could also have abstract protected functions as well. Which one you use just depends on your use case. It's a very powerful feature of object oriented code.