Here's the really beautiful thing about abstract classes. You may create some of these because you have a situation similar to the one we've been working on in this project. Or, you may be using someone else's code like a third party library that you downloaded via the Composer package manager. You might even read in that library's documentation that if you want to create a new ship class you just need to extend AbstractShip.
What's really great is that AbstractShip now tells you exactly what you need to do to create a new ship class with its three abstract functions that you must fill in:

 123 lines  lib/Model/AbstractShip.php
... lines 1 - 2
abstract class AbstractShip
{
... lines 5 - 15
    abstract public function getJediFactor();
... lines 17 - 20
    abstract public function getType();
... lines 22 - 25
    abstract public function isFunctional();
... lines 27 - 121
}
A third group has joined the battle and we have a new type of ship. They're not very good mechanics, so we'll call this a broken ship. This is simple, the ship is always broken.