A third group has joined the battle and we have a new type of ship. They're not very good mechanics, so we'll call this a broken ship. This is simple, the ship is always broken.
Create a new php class called BrokenShip. Of course now make it extend AbstractShip.

Let's pretend like we don't know that there are any abstract methods in the parent class. So we won't do anything here except putting in the extends code. Head over to bootstrap.php and require our useless new BrokenShip.

Back in index.php for now, let's just add {dollar_sign}brokenShip = new BrokenShip(); and add it to our ships array.

We can do this because we know that BrokenShip extends AbstractShip. And down here, when we use those ship objects we're just calling methods on the AbstractShip.

Back to the browser, refresh! Yes, what a huge beautiful error. It says:

Class BrokenShip contains 3 abstract methods and must therefore be declared abstract or implement the remaining methods.
And then it goes on and lists the methods.

And then it goes on and lists the methods.

In other words, it's saying 'Hey buddy! You need to add those three methods into this class! ' It's always giving you an out to declare the class abstract if you want to, and you might do this if you wanted an abstract class inside an abstract class with some additional public functions. But we've all seen where that goes in the move inception.