To get our ships we use ShipLoader which queries the database and creates ship objects. This queryForShips() goes out, selects all the ships, and then later it is passed to this nice createShipFromData() function down here:
 61 lines  lib/Service/ShipLoader.php
... lines 1 - 2
class ShipLoader
{
... lines 5 - 14
    public function getShips()
    {
      {dollar_sign}ships = array();
      {dollar_sign}shipsData = {dollar_sign}this->queryForShips();
      foreach ({dollar_sign}shipsData as {dollar_sign}shipData) {
        {dollar_sign}ships[] = {dollar_sign}this->createShipFromData({dollar_sign}shipData);
      }
      return {dollar_sign}ships;
    }
... lines 27 - 58
}
... lines 60 - 61
This is the one we've been working in that creates the objects.

Step 1: Query the database
Step 2: Turn that data into objects
Suppose that we have a new requirement, sometimes we're going to get the ship data from the database but other times it will come from a different source, like a JSON file.

In the resources directory there's a new ship.json file, as you can see this holds the same info as we have in the database:

 35 lines  resources/ships.json
... line 1
[
    {
      'id': '1',
      'name': 'Jedi Starfighter',
      'weapon_power': '5',
      'jedi_factor': '15',
      'strength': '30',
      'team': 'rebel'
    },
... lines 11 - 26
    {
      'id': '4',
      'name': 'RZ-1 A-wing interceptor',
      'weapon_power': '4',
      'jedi_factor': '4',
      'strength': '50',
      'team': 'empire'
    }
]
Now why would we want our application to sometimes load from the database and other times from a JSON file? Say that when we're developing locally we don't have access to our database, so we use a JSON file. But when we push to production we'll switch back to the real database. Or, suppose that our ship library is so awesome that someone else wants to reuse it. However, this fan doesn't use a database, they only load them from JSON.

This leaves us needing to make our ShipLoader more generic.