Create a new class whose only job is to load ship data through the database, or PDO.
Looking back inside ShipLoader there are two types of queries that we make. Sometimes we query for all of the ships and sometimes we query for just one ship by ID.

Create two methods, to cover both of those actions. First, create a public function fetchAllShipsData() which we'll fill out in just one second. Now, add public function fetchSingleShipData() and pass it the id that we want to query for.

Before we go any further head back to our bootstrap.php file and make sure that we require this.