Head over to bootstrap of course and require JsonFileShipStorage.php.
In theory since this class has all the same methods as PdoShipStorage we should be able to pass a JsonFileShipStorage object into ShipLoader and everything should just work. Really, the only thing ShipLoader should care about is that it's passed an object that has the two methods it's calling: fetchAllShipsData() and fetchSingleShipData():
61 lines  lib/Service/ShipLoader.php
... lines 1 - 2
class ShipLoader
{
... lines 5 - 31
  public function findOneById({dollar_sign}id){
    {dollar_sign}shipArray = {dollar_sign}this->shipStorage->fetchSingleShipData({dollar_sign}id);
... lines 35 - 36
  }
... lines 38 - 54
  private function queryForShips(){
    return {dollar_sign}this->shipStorage->fetchAllShipsData();
  }
}

In Container let's give this a try. Down in getShipStorage() let's say, {dollar_sign}this->shipStorage = new JsonFileShipStorage(). And we'll give it a path to our JSON of __DIR__.'/../../resources/ships.json'

From this directory I'm going up a couple of levels, into resources and pointing at this ships.json file which holds all of our ship info.

Back to the browser and refresh. Ok no success yet, but as they say, try try again. Before we do that, let's check out this error:

Argument 1 passed to ShipLoader::__construct() must be an instance of PdoShipStorage, instance of JsonFileShipStorage given.
What's happening here is that in ShipLoader we have this type-hint which says that we only accept PdoShipStorage and our Json file is not an instance of that:

 61 lines  lib/Service/ShipLoader.php
... lines 1 - 2
class ShipLoader {
... lines 5 - 6
  public function __construct(PdoShipStorage ){
... line 9
  }
... lines 11 - 58
}