The easiest way to fix this is to say extends PdoShipStorage in JsonFileShipStorage.
This makes the json file an instance of PdoShipStorage. Try refreshing that again. Perfect, our site is working.

But having to put that extends in our JSON file kinda sucks, when we do this we're overriding every single method and getting some extra stuff that we aren't going to use.