You should be thinking, 'This is a good spot for Abstract Ship Storage! ' And well, I agree so let's create that. Inside the Service directory add a new PHP Class called AbstractShipStorage. The two methods that this is going to need to have are: fetchSingleShipData() and fetchAllShipsData() so I'll copy both of those and paste them over to our new class.
Of course we don't have any body in these methods, so remove that. Now, set this as an abstract class. Make both of the functions abstract as well.

Cool!

Now, JsonFileShipStorage can extend AbstractShipStorage.
And the same thing for PdoShipStorage.

With this setup we know that if we have a AbstractShipStorage it will definitely have both of those methods so we can go into the ShipLoader and change this type hint to AbstractShipStorage because we don't care which of the two storage classes are actually passed.

To be very well behaved developers, we'll go into our Container and above getShipStorage() change the type hint to AbstractShipStorage. Not a requirement, but it is a good idea.

Go back to the browser and refresh... oh, class AbstractShipStorage not found because we forgot to require it in our bootstrap file. We will eventually fix the need to have all of these require statements.

Refresh again and now it works perfectly.

We created an AbstractShipStorage because it allows us to make our ShipLoader more generic. It now doesn't care which one is passed, as long as it extends AbstractShipStorage.

But there's an even better way to handle this... interfaces!