Notice, AbstractShipStorage unlike AbstractShip, doesn't actually have any logic in it:
 8 lines  lib/Service/AbstractShipStorage.php
... lines 1 - 2
abstract class AbstractShipStorage
{
  abstract public function fetchAllShipsData();
  abstract public function fetchSingleShipData({dollar_sign}id);
}
All it does is have a contract that guarantees anything that extends this has these two functions. It turns out that when you have an abstract class like this that only contains abstract functions and no real code, well it's the perfect opportunity to use an Interface.

An interface works just like an abstract class and here's how it looks. To start, we need to rename our class to ShipStorageInterface since this more closely matches what it is. And instead of abstract class it's now labeled as an interface.

Get it?

As soon as you do that you no longer need abstract in front of all the functions, but these work the same.

On the AbstractShipStorage file in the tree, go to 'Refactor' and click to Rename our file to ShipStorageInterface. I really like the consistency. And of course update our require line for this file in bootstrap.php.