What I want to do is move all the querying logic from ShipLoader into this PdoShipStorage class. Let's start with the logic that queries for one ship and pasting that over.
Notice, that we're not returning an object here this is just a really dumb class that returns data, an array in our case.

There is one problem, we have a getPdo() function inside of ShipLoader that references a pdo property. Point being, our PDO storage needs access to the PDO object, so we're going to use dependency injection, a topic we covered a lot in episode 2 . Add public function __construct(PDO {dollar_sign}pdo) and store it as a property with {dollar_sign}this->pdo = {dollar_sign}pdo.

If this pattern is new to you just head back and watch the dependency injection video in episode 2 of the OO series.

Here we're saying, whomever creates our PDO ship storage class must pass in the pdo object. This is cool because we need it. Now I can just reference the property there directly.