Back in ShipLoader copy the entire queryForShips() and paste that into fetchAllShipsData() and once again reference the pdo property.
Now we have a class whose only job is to query for ship stuff, we're not using it anywhere yet, but it's fully ready to go. So let's use this inside of ShipLoader instead of the PDO object. Since we don't need PDO to be passed anymore swap that out for a PdoShipStorage object. Let's update that in a few other places and change the property to be called shipStorage.

Cool!