In getShips() we used to call ->queryForShips(); but we don't need to do that anymore!
Perfect, now scroll down and get rid of the queryForShips() function all together: we're not using that anymore. And while we're cleaning things out also delete this getPDO() function. We can delete this because up here where we reference it in findOneById() we'll do the same thing. Remove all the pdo querying logic, and instead say shipArray = {dollar_sign}this->shipStorage->fetchSingleShipData(); and pass it the ID.

This class now has no query logic anywhere.