All we know is that we're passed in some PdoShipStorage object and we're able to call methods on it. It can make the queries and talk to whatever database it wants to, that's it's responsibility. In here we're just calling methods instead of actually querying for things.
ShipLoader and PdoShipStorage are now fully setup and functional. The last step is going into our container which is responsible for creating all of our objects to make a couple of changes. For example, when we have new ShipLoader we don't want to pass a pdo object anymore we want to pass in PdoShipStorage.

Just like before, create a new function called getShipStorage() and make sure we have our property up above. The getShipStorage() method is going to do exactly what you expect it to do. Instantiate a new PdoShipStorage and return it. The ship's storage class does need PDO as its first constructor argument which we do with new PdoShipStorage(->getPDO())