Everything used to be in ShipLoader, including the query logic. We've now split things up so that the query logic is in PdoShipStorage and in ShipLoader you're just calling methods on the shipStorage. Its real job is to create the objects from the data, wherever that data came from. In Container.php we've wired all this stuff up.
Phew, that was a lot of coding we just did, but when we go to the browser and refresh, everything still works exactly the same as before. That was a lot of internal refactoring. In index.php as always we still have {dollar_sign}shipLoader->getShips().
And that function still works as it did before, but the logic is now separated into two pieces.

The cool thing about this is that our classes are now more focused and broken into smaller pieces. Initially we didn't need to do this, but once we had the new requirement of needing to load ships from a JSON file this refactoring became necessary. Now let's see how to actually load things from JSON instead of PDO.